Initial commit
This is straight from the Hosting Reborn website. Well, nearly:

- The SCSS is now compacted, not compressed (for easier debug)
  and stored in `/style.css`, not `/css/style.css`.

- The JS is manually cat'd using
  `cat \`find js/ | sort -r | xargs\` > script.js` instead of this
  being done at request-time by server-side includes.

The apple icons are not really necessary IMO, but I'll leave them
there until I have investigated more.

While on that topic, the favicon needs to be redone as it still is
the old white/black one, even though the website changed to a grey
design. Or maybe it's all right?