Replace e with é in my name.
Félix is my proper name. I was using Felix (without accent) on most
services until about a month ago, when I decided to do the switch.
This does mean my name is less easy to type on a normal layout (thus
many people will continue to type my name as Felix), but this is in
alignment of my web indentity shift of the past few months (from a
linked passcod/personal to a unique, fully merged, identity across
web, paper, and meat-space.